+++ Client/Client.Commands.cs
@@ -460,8 +460,34 @@ public partial class Client
         }
     }
 
+    [ConsoleCommand("nextmap", "Starts a new game on the next map")]
+    private void CommandNextMap(ConsoleCommandEventArgs args)
+    {
+        var maps = m_archiveCollection.MapInfo.MapInfo.GetOrderedMaps();
+        if (maps.Count == 0)
+            return;
+        int nextIndex = (m_layerManager.WorldLayer != null)
+            ? (maps.IndexOf(m_layerManager.WorldLayer.CurrentMap) + 1) % maps.Count
+            : 0;
+        MapInfoDef map = maps[nextIndex];
+        NewGame(map);
+    }
+
+    [ConsoleCommand("previousmap", "Starts a new game on the previous map")]
+    private void CommandPreviousMap(ConsoleCommandEventArgs args)
+    {
+        var maps = m_archiveCollection.MapInfo.MapInfo.GetOrderedMaps();
+        if (maps.Count == 0)
+            return;
+        int prevIndex = (m_layerManager.WorldLayer != null)
+            ? (Math.Max(0, maps.IndexOf(m_layerManager.WorldLayer.CurrentMap)) - 1 + maps.Count) % maps.Count
+            : maps.Count - 1;
+        MapInfoDef map = maps[prevIndex];
+        NewGame(map);
+    }
+
     [ConsoleCommand("listmaps", "Lists all maps")]
-    private void ListMaps(ConsoleCommandEventArgs args)
+    private void CommandListMaps(ConsoleCommandEventArgs args)
     {
         foreach (var map in m_archiveCollection.MapInfo.MapInfo.Maps)
         {
@@ -476,13 +502,13 @@ public partial class Client
     }
 
     [ConsoleCommand("printmap", "Prints the current map")]
-    private void PrintMap(ConsoleCommandEventArgs args)
+    private void CommandPrintMap(ConsoleCommandEventArgs args)
     {
         HelionLog.Info(GetCurrentMapName() ?? "No map loaded");
     }
 
     [ConsoleCommand("printgame", "Prints the current game title (or WAD filename)")]
-    private void PrintGame(ConsoleCommandEventArgs args)
+    private void CommandPrintGame(ConsoleCommandEventArgs args)
     {
         HelionLog.Info(GetCurrentGameName() ?? "No map loaded");
     }